requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
mcp[cli]>=1.27.0,<2
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
anyio>=4.4.0
starlette>=0.40.0
sse-starlette>=2.1.0
